+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/execution/config/FileUtils.java
@@ -4,7 +4,7 @@ import org.jetbrains.annotations.NotNull;
 
 import java.nio.file.Path;
 
-public final class FileUtils {
+final class FileUtils {
 
     private FileUtils() {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");